url shortener. After review refactoring Interfaces added to repositories
