+++ src/main/java/faang/school/urlshortenerservice/dto/ShortUrlHashDto.java
@@ -1,0 +1,8 @@
+package faang.school.urlshortenerservice.dto;
+
+import lombok.Data;
+
+@Data
+public class ShortUrlHashDto {
+    private String urlHash;
+}
