+++ src/main/java/faang/school/urlshortenerservice/repository/CacheRepository.java
@@ -1,0 +1,8 @@
+package faang.school.urlshortenerservice.repository;
+
+import java.util.Optional;
+
+public interface CacheRepository {
+    void put(String key, String value);
+    Optional<String> get(String key);
+}
