+++ src/main/java/faang/school/urlshortenerservice/repository/CacheRepositoryRedisImpl.java
@@ -9,15 +9,17 @@ import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
-public class RedisRepository {
+public class CacheRepositoryRedisImpl implements CacheRepository {
 
     private final RedisTemplate<String, String> redisTemplate;
     private final Duration cacheTtl;
 
+    @Override
     public void put(String key, String value) {
         redisTemplate.opsForValue().set(key, value, cacheTtl);
     }
 
+    @Override
     public Optional<String> get(String key) {
         String value = redisTemplate.opsForValue().get(key);
         return Optional.ofNullable(value);
