+++ src/main/java/faang/school/urlshortenerservice/service/UrlServiceImpl.java
@@ -2,8 +2,8 @@ package faang.school.urlshortenerservice.service;
 
 import faang.school.urlshortenerservice.hash.HashCache;
 import faang.school.urlshortenerservice.config.HashEncoderProperties;
-import faang.school.urlshortenerservice.repository.RedisRepository;
-import faang.school.urlshortenerservice.repository.UrlRepository;
+import faang.school.urlshortenerservice.repository.CacheRepositoryRedisImpl;
+import faang.school.urlshortenerservice.repository.UrlRepositoryJdbcImpl;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,8 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class UrlServiceImpl implements UrlService {
-    private final UrlRepository urlRepository;
-    private final RedisRepository redisRepository;
+    private final UrlRepositoryJdbcImpl urlRepository;
+    private final CacheRepositoryRedisImpl cacheRepository;
     private final HashCache hashCache;
     private final HashEncoderProperties hashEncoderProperties;
 
@@ -22,19 +22,19 @@ public class UrlServiceImpl implements UrlService {
     public String shortAndReturn(String longUrl) {
         String hash = hashCache.getHash();
         urlRepository.save(hash, longUrl);
-        redisRepository.put(hash, longUrl);
+        cacheRepository.put(hash, longUrl);
         return hash;
     }
 
     @Override
     public String findOriginal(String hash) {
         validateHash(hash);
-        return redisRepository
+        return cacheRepository
                 .get(hash)
                 .orElseGet(() -> urlRepository
                         .findByHash(hash)
                         .map(longUrl -> {
-                            redisRepository.put(hash, longUrl);
+                            cacheRepository.put(hash, longUrl);
                             return longUrl;
                         })
                         .orElseThrow(EntityNotFoundException::new));
